+++ CHANGES.md
@@ -20,6 +20,7 @@ Changes in 8.0
 - Remove Waived Fees from list in Student_Billing/functions.inc.php
 - New ROSARIO_DISABLE_ADDON_DELETE optional config constant in INSTALL.md, INSTALL_es.md & INSTALL_fr.md
 - Add-on disable delete in Modules.inc.php & Plugins.inc.php
+- Merge Daily Transactions & Daily Totals programs in DailyTransactions.php & DailyTotals.php
 
 Changes in 7.9.3
 ----------------